:truck: refactor: 패키지로 디렉토리 변경 (#28)
