+++ src/module/family-member.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import {
   FamilyMemberController,
   FamilyMemberService,
-} from '../domain/family-member';
+} from '../domain/family_member';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Family, FamilyMember, User } from '../infra/entities';
